// shared/types.ts
export interface Customer {
  id:string;
  name: string;
  cpfCnpj: string;
  email: string;
  phone: string;
  address?: string; // Endereço para entregas
}

// Representa um item em uma venda
export interface SaleItem {
  productId: string;
  productName: string;
  quantity: number;
  returnableQuantity: number; // Quantidade que ainda pode ser devolvida
  unitPrice: number;
  totalPrice: number;
  imageUrl?: string;
}

// Representa uma venda completa
export interface Sale {
  id:string;
  date: string; // ISO string format
  items: SaleItem[];
  totalAmount: number;
  paymentMethod: 'Dinheiro' | 'Cartão de Crédito' | 'Cartão de Débito' | 'PIX' | 'A Pagar na Entrega' | 'Troca / Vale-Crédito';
  customerId?: string;
  customerName?: string;
  deliveryId?: string; // ID da entrega associada
  status: 'Completed' | 'Partially Returned' | 'Fully Returned' | 'Pending Payment';
  storeCreditAmountUsed?: number; // Valor pago com vale-crédito
  // Dados da NFC-e
  nfceAccessKey?: string;
  nfceQrCodeUrl?: string;
}
// ...
export interface StoreCredit {
  id: string;
  customerId: string;
  customerName: string;
  initialAmount: number;
  balance: number;
  createdAt: string; // ISO string
  expiresAt: string; // ISO string
  status: 'Active' | 'Used';
}